requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
openai==1.3.5
python-dotenv==1.0.1
youtube-transcript-api==0.6.1
pydantic==2.5.2
